Swap out the domain for the looked up IP address.
hit looks up the IP address so it can display it however hyper was doing it's
own lookups. This meant the displayed IP address may have been different to that
hit in the case of round robin DNS for example. This commit fixes that by
swapping out the domain for the IP address in the URL. A host header was already
being set. Not swapping the domain for the IP address was an oversight.